requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
prisma==0.11.0
pytest==7.4.3
python-multipart==0.0.6
fastapi
uvicorn
httpx[http2]
pytest
pytest-asyncio
